test: move package unit tests into package-level tests directories
Unit tests move out of src/ (and __tests__/) into each package's tests/
directory, mirroring the src/ structure. Package tsconfigs include tests/
so pnpm check type-checks them — which is the entire assertion for the
types.test.ts suites. The ported react-docgen-typescript __tests__ suite
stays in its upstream layout.
