+++ packages/addon-rslib/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {},
-  "include": ["src/**/*"]
+  "include": ["src/**/*", "tests/**/*"]
 }
