+++ packages/builder-rsbuild/tests/types.test.ts
@@ -1,13 +1,13 @@
 import { describe, expect, it } from '@rstest/core'
-import type { StorybookConfig } from './types'
+import type { StorybookConfigRsbuild } from '../src/types'
 
 const features = {
   experimentalTestSyntax: true,
   babelRemoveBugfixes: true,
-} satisfies NonNullable<StorybookConfig['features']>
+} satisfies NonNullable<StorybookConfigRsbuild['features']>
 
-describe('StorybookConfig', () => {
-  it('accepts React framework feature flags', () => {
+describe('StorybookConfigRsbuild', () => {
+  it('accepts builder feature flags', () => {
     expect(features).toEqual({
       experimentalTestSyntax: true,
       babelRemoveBugfixes: true,
