+++ packages/framework-react/tests/loaders/tsconfig-paths.test.ts
@@ -2,7 +2,7 @@ import { mkdirSync, mkdtempSync, rmSync, writeFileSync } from 'node:fs'
 import { tmpdir } from 'node:os'
 import { dirname, join } from 'node:path'
 import { afterEach, describe, expect, it } from '@rstest/core'
-import { getTsconfigPathsBaseDir } from './tsconfig-paths'
+import { getTsconfigPathsBaseDir } from '../../src/loaders/tsconfig-paths'
 
 const tempDirs: string[] = []
