+++ packages/framework-react/tests/preset.test.ts
@@ -1,7 +1,7 @@
 import { createRequire } from 'node:module'
 import type { RsbuildConfig } from '@rsbuild/core'
 import { describe, expect, it } from '@rstest/core'
-import { rsbuildFinal } from './preset'
+import { rsbuildFinal } from '../src/preset'
 
 const require = createRequire(import.meta.url)
